Authorization Notification and changes to the Notification System
Added new field(invalid_credentials) to emailNotification that get set when invalid crednetials are used to get mail or send mail.
Updated the AutoUpdater to toast notifications if the app is in the foreground, and use notifications when the app is in the background.
Added validation for credentials when leaving from login activity.
Changed  getContactName in Database Helper to give full name instead of only first name.
Added gradle for Xdroid Toaster for multithreaded toasting. This was included from https://github.com/shamanland/xdroid