editContact
created the editContact Activity and class
This class allows for the contacts information to be changed and still retain the conversation data.(see DataBaseHelper updateConversationWindowWithDifferentEmail for more details)
Removed activity_fake_service.xml. It was not being used. Git notes this as a file change to activity_edit_contact.xml.
I also moved this from onClick to onLongClick