Multiple Emails
Implemented a system that can use one of a few diffent emails. Gmail, Yahoo, Outlook have been tested. AOL, live, icloud and hotmail were also added but were not tested.
Implemented chaged to the HelperClass, which needs renamed. njw was right. The class did com in handy.
Created Send Mail class. I brought the class out from as an inner class just in case it is needed elsewhere. My current thoughts are that we may need it for the QR feature with a new subscribe feature.
Updated contactFragment and conversationFragment to have more appropriate toasts.