Intermittent Push
Added CM.minimize(Textview) to minimize textViews when not used. (Also there is an expand function.)
Added a check to make sure that attachments are not downloaded twice. Will need refactored to simplify.
Several known bugs, that will get fixed in the next push. There are some issuse with working with the contact list and the contact list in getMail. There is also an issue with the catch block not being close enough and causing issues in getMail.