Major refactoring with Contacts and Conversations
The goal of the refactor was to fix the issue of when a conversation was not started but there was a new email for the contact.
Changed the data base to better incorporate the data needed.
Moved createdDate and UpdatedDate from Conversation to Contact.
Deleted Notification database and put it in the contact database.
Changed ConversationWindow to Message. Although this may bring some confusion with the mail Message, i believe it to be a more appropraite name.