Changes to Deletion and other improvements
changed deletion of contacts which cause an error that would permantly disable the app and it would need a delete and install.
refactored the detetion of Conversations to delete messages and attachments as well.
changed the insert of contacts to not add a conversation exception from the add in main activity.
Added a length maximun for notifications in getMail.
Replaced depreciated functions in getContactsFromInbox in getMail.
Edited the documentation docx.