Sent Messages
Added sent messages feature
-Created new section of getMail() to get messages from the sent box to the given email.
-Added a few lines of code to make the created date for new contacts back two weeks from the creation point so it pulls in the mail.
Fixed duplication in Settings
Changed layout for messages so subject could be neatly displayed/bolded
Replaced gotMail from email Notifcation with a call to size;
Removed Invalid_Credentials because it is not longer needed with the refactoring of LoginActivity
Replaced emailNotification with a simple Stack<NotificationBase> because there is no reason to complicated the code anymore.
Added padding to the messages would not be quite so on top of each other via the java code so it is not added for times.
Updated Documentation
Added feature to remove sent_from_app messages with thier server counter-part